backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
sqlalchemy>=2.0.0
plaid-python>=16.0.0
python-dotenv>=1.0.0

ruff==0.15.2